move classes to model
